requirements.txt
attrdict==2.0.0
certifi==2018.10.15
chardet==3.0.4
cytoolz==0.9.0.1
Django==2.1.2
eth-abi==1.2.2
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-typing==1.3.0
eth-utils==1.2.2
hexbytes==0.1.0
idna==2.7
logzero==1.5.0
lru-dict==1.1.6
parsimonious==0.8.1
psycopg2==2.7.5
py-solc==3.2.0
pycryptodome==3.7.0
pytz==2018.7
requests==2.20.0
rlp==1.0.3
semantic-version==2.6.0
six==1.11.0
toolz==0.9.0
urllib3==1.24
web3==4.8.1
websockets==6.0
